requirements.txt
beautifulsoup4==4.11.1
Django==4.0.1
django-bootstrap-v5==1.0.11
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-webpack-loader==1.4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
entrypoints==0.4
executing==0.8.2
google-search-results==2.4.1
gunicorn==20.1.0
idna==3.3
ipykernel==6.9.1
ipython==8.0.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
oauthlib==3.1.1
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
stripe==2.65.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
tzdata==2022.1
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==5.3.0
